refactor: file hierarchy
